Add files via upload
This project looks at a variety of data in order to identify the data that may be correlated to overall airline delays. By comparing airline delays to arrival delays, departure delays, specific states, airports, and months in the year, as well as air time and travel distance, the project works to identify any characteristics that appear to have a strong relationship with airline delay within this data set. After analyzing the data, it becomes clear that a variety of factors appear  to be correlated with airline delay, including arrival and departure delays, longer distances, and greater airtime. However, the factors with the largest influence on higher than average airline delays seem to be the month of the year and the location of flights. Although this insight  cannot be explicitly drawn from the available data, it appears that the human factor (perhaps surrounding holidays and travel as well as location) may be the greatest influence on larger than average airline delays.